+++ R/xrd.get.PeakStats.R
@@ -1,11 +1,14 @@
-#' Fits the peak to a Gaussian
+#' Fits a Gaussian Function to XRD data
 #'
+#' @description
 #' usually run after a peak is
 #' identified through its location
 #'
-#' @param TwoTheta angle
-#' @param Intensity intensity signal
-#' @param PeakPos approximate angle of peak position
+#' @author Thomas Gredig
+#'
+#' @param TwoTheta 2q angle
+#' @param Intensity XRD intensity
+#' @param peakPos specific 2q angle of peak position
 #' @param Try.Sigma vector with peak widths used to start fitting
 #' @param verbose logical, if \code{TRUE} output additional information
 #'
@@ -15,41 +18,54 @@
 #' d = xrd.read.ASC(filename)
 #' xrd.get.PeakStats(d$theta, d$I, 38.2)
 #'
-#' @importFrom stats nls sd
+#' @importFrom stats nls sd predict
+#' @importFrom graphics lines
 #' @export
-xrd.get.PeakStats <- function(TwoTheta, Intensity, PeakPos,
-                              Try.Sigma = c(0.1,0.4,0.2,0.15),
+xrd.get.PeakStats <- function(TwoTheta, Intensity, peakPos,
+                              Try.Sigma = c(0.2,0.1,0.05,0.3),
                               verbose = FALSE) {
   d = data.frame(TwoTheta, I = Intensity)
-  for(peak.width in Try.Sigma) {
+  for(pw in Try.Sigma) {
+
+    deltaPeak = max(0.2, 4*pw)
     fit <- NULL
-    n1 = subset(d, TwoTheta > (PeakPos-2.5*peak.width) &
-                  TwoTheta < (PeakPos+2.5*peak.width))
-    if (verbose) print(paste("Set from 2q =",
+    n1 = subset(d, TwoTheta > (peakPos-deltaPeak) &
+                  TwoTheta < (peakPos+deltaPeak))
+    if (verbose) plot(n1)
+    if (verbose) print(paste("Fit from 2q =",
                              min(n1$TwoTheta), "to",
                              max(n1$TwoTheta)))
     if(nrow(n1)>5) {
-      p1 = n1$TwoTheta[which.max(n1$I)]
-      background = min(n1$I)
-      A1 = max(n1$I) - background
-      if (A1 > (3*sd(n1$I))) {
-        if (verbose) {
-          print(paste("background=",background,
-                      " A=",A1,
-                      " t0=",p1,
-                      " sigma=0.2"))
-        }
-        try(fit <-
-              nls(data = n1,
-                  I ~ b + A*exp(-(TwoTheta-t0)^2/(2*sigma*sigma)),
-                  start = list(b=background, A = A1, t0=p1, sigma=0.2))
-        ); # does not stop in case of error
-      } else {
-        fit = NULL
+      q = xrd.peakEstimate(n1$TwoTheta, n1$I)
+      background = ceiling(q$b0)
+      A1 = q$A0
+      p1 = q$th0
+      peak.width = pw
+
+      if (verbose) {
+        print(paste("background=",background,
+                    " A=",A1,
+                    " t0=",p1,
+                    " sigma=",peak.width))
       }
+      # if ((min(n1$TwoTheta)==p1) | (min(n1$TwoTheta)==p1)) break;
+      # if (sd(n1$I)*4 > A1) break;
+      try(fit <-
+            nls(data = n1,
+                I ~ b + A*exp(-(TwoTheta-th1)^2/(2*sigma*sigma)),
+                start = list(b=background, A = A1, th1=p1, sigma=pw))
+      ); # does not stop in case of error
+
       if (!is.null(fit)) break;
     }
   }
+  if (verbose) {
+    if (!is.null(fit)) {
+      plot(n1)
+      lines(n1$TwoTheta, predict(fit),col='red')
+    }
+  }
+
   if(is.null(fit)) { return(NA) }
   summary(fit)$coeff[1:8]
 }