+++ tests/testthat/test-DebyeScherrer.R
@@ -1,8 +1,8 @@
 test_that("get Debye-Scherrer width", {
   fn = xrd.getSampleFiles(fileExt='asc')
   d <- xrd.read.ASC(fn)
-  q = xrd.get.PeakStats(d$theta, d$I, 38.2)
+  q = xrd.get.PeakStats(d$theta, d$I, 38.2, Try.Sigma = 0.3)
 
-  # expect Au layer to be 14.7nm
-  expect_equivalent(xrd.get.DebyeScherrer(q), 152.2, tolerance=1e-3)
+  # expect Au layer to be 14.4nm
+  expect_equal(xrd.get.DebyeScherrer(q), 144, tolerance=1e-3)
 })